Model or Entity fix; author and copyrighttags addes; composer.json: deleted some dependencies
