+++ Model/SubscriptionWall.php
@@ -1,12 +1,21 @@
 <?php
 
-namespace Integrated\Bundle\SubscriptionBundle\Entity;
+/*
+ * This file is part of the Integrated package.
+ *
+ * (c) e-Active B.V. <integrated@e-active.nl>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Integrated\Bundle\SubscriptionBundle\Model;
 
 use Doctrine\ORM\Mapping as ORM;
 use Doctrine\Common\Collections\ArrayCollection;
 
 /**
- * Integrated\Bundle\SubscriptionBundle\Entity\SubscriptionWall
+ * @author Jacob de Graaf <jacob.de.graaf@windesheim.nl> and Albert Bakker <albert-david.bakker@windesheim.nl>
  */
 class SubscriptionWall
 {
@@ -31,12 +40,20 @@ class SubscriptionWall
     protected $disabled;
 
     /**
+     * Times a user can see an article before payment
+     *
      * @var string
      */
-    protected $freetier;
+    protected $freeTier;
 
+    /**
+     * @var WallChannel[]
+     */
     protected $wallChannels;
 
+    /**
+     * @var SubscriptionType[]
+     */
     protected $subscriptionTypes;
 
     public function __construct()
@@ -47,7 +64,7 @@ class SubscriptionWall
 
     /**
      * @param string $id
-     * @return \Integrated\Bundle\SubscriptionBundle\Entity\SubscriptionWall
+     * @return \Integrated\Bundle\SubscriptionBundle\Model\SubscriptionWall
      */
     public function setId($id)
     {
