+++ Model/VatType.php
@@ -1,12 +1,21 @@
 <?php
 
-namespace Integrated\Bundle\SubscriptionBundle\Entity;
+/*
+ * This file is part of the Integrated package.
+ *
+ * (c) e-Active B.V. <integrated@e-active.nl>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Integrated\Bundle\SubscriptionBundle\Model;
 
 use Doctrine\ORM\Mapping as ORM;
 use Doctrine\Common\Collections\ArrayCollection;
 
 /**
- * Integrated\Bundle\SubscriptionBundle\Entity\VatType
+ * @author Jacob de Graaf <jacob.de.graaf@windesheim.nl> and Albert Bakker <albert-david.bakker@windesheim.nl>
  */
 class VatType
 {
@@ -45,10 +54,19 @@ class VatType
      */
     protected $continent;
 
+    /**
+     * @var SubscriptionType[]
+     */
     protected $subscriptionTypes;
 
+    /**
+     * @var VatCountry
+     */
     protected $vatCountry;
 
+    /**
+     * @var VatContinent
+     */
     protected $vatContinent;
 
     public function __construct()
