+++ Resources/config/mapping/doctrine/SubscriptionType.orm.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-  <entity name="Integrated\Bundle\SubscriptionBundle\Entity\SubscriptionType" table="subscription_type">
+  <entity name="Integrated\Bundle\SubscriptionBundle\Model\SubscriptionType" table="subscription_type">
 
     <indexes>
       <index columns="type_vat"/>
