+++ Resources/config/mapping/doctrine/SubscriptionWall.orm.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-  <entity name="Integrated\Bundle\SubscriptionBundle\Entity\SubscriptionWall" table="subscription_wall">
+  <entity name="Integrated\Bundle\SubscriptionBundle\Model\SubscriptionWall" table="subscription_wall">
 
     <id name="id" type="string" column="id" length="36">
       <generator strategy="UUID"/>
@@ -21,7 +21,7 @@
 
     <field name="disabled" type="boolean" column="disabled" nullable="true"/>
 
-    <field name="freetier" type="string" column="freetier" length="45" nullable="true">
+    <field name="freeTier" type="string" column="freetier" length="45" nullable="true">
       <options>
         <option name="fixed"/>
       </options>
