+++ Resources/config/mapping/doctrine/VatContinent.orm.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-  <entity name="Integrated\Bundle\SubscriptionBundle\Entity\VatContinent" table="vat_continent">
+  <entity name="Integrated\Bundle\SubscriptionBundle\Model\VatContinent" table="vat_continent">
 
     <id name="continentCode" type="string" column="continent_code" length="2">
       <generator strategy="IDENTITY"/>
