+++ Resources/config/mapping/doctrine/VatCountry.orm.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-  <entity name="Integrated\Bundle\SubscriptionBundle\Entity\VatCountry" table="vat_country">
+  <entity name="Integrated\Bundle\SubscriptionBundle\Model\VatCountry" table="vat_country">
 
     <id name="countryCode" type="string" column="country_code" length="3">
       <generator strategy="IDENTITY"/>
