+++ composer.json
@@ -18,15 +18,13 @@
         "symfony/symfony": ">=2.4",
         "symfony/framework-bundle": ">=2.2",
         "sensio/framework-extra-bundle": "~2.3",
-        "doctrine/mongodb": ">=1.0.0",
-        "doctrine/mongodb-odm-bundle": "~3.0",
-        "integrated/content-bundle": "~0.5",
-        "knplabs/knp-paginator-bundle": ">=2.4"
+        "integrated/content-bundle": "~0.5"
     },
     "autoload":{
         "psr-4":{
             "Integrated\\Bundle\\SubscriptionBundle\\": ""
         }
+
     },
     "extra": {
         "branch-alias": {
